# docs/openapi.yaml
...
openapi: 3.0.3
...
info:
  title: Hajj Permit API
  version: 1.0.0
  description: >
    Backend API for the Hajj Permit course project.
    It provides entity management, inventory requests,
    inventory management, and vehicle permit issuance.

# Clients access the system through the Backend Gateway.
servers:
  - url: http://localhost:3000
    description: Local Docker environment through Backend Gateway

# Groups used to organize the API documentation.
...
paths:
...
  /entities:
    get:
      tags:
        - Entity
      operationId: getEntities
      summary: Get all entities
      responses:
        "200":
          description: List of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"

    post:
      tags:
        - Entity
      operationId: createEntity
      summary: Create a new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
              properties:
                id:
                  type: string
                  example: E3
                name:
                  type: string
                  example: Riyadh Transport
                active:
                  type: boolean
                  default: true
                  example: true

      responses:
        "201":
          description: Entity created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"

        "400":
          description: Required information is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

              example:
                message: id and name are required

        "409":
          description: Entity already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

              example:
                message: Entity already exists
  /entities/{id}:
    get:
      tags:
        - Entity
      operationId: getEntityById
      summary: Get one entity by ID

      parameters:
        - name: id
          in: path
          required: true
          description: Entity ID
          schema:
            type: string
          example: E1

      responses:
        "200":
          description: Entity was found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"

        "404":
          description: Entity was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

              example:
                message: Entity not found


  # ---------------------------------------------------------
  # Inventory
  # ---------------------------------------------------------
...
components:
  schemas:
    Entity:
      type: object
      required:
        - id
        - name
        - active
      properties:
        id:
          type: string
          example: E1

        name:
          type: string
          example: Alpha Transport

        active:
          type: boolean
          description: Whether the entity is active and eligible
          example: true
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Entity not found